refactor: Remove extraneous sources
These sources were added in the first place because each file
that contained them was exec'd. This means that it did not share the
sourced functions that the parent Bash program already sourced. However,
since these exec's have been removed, the functions are now shared and no longer
need to be re-sourced