refactor: Remove 'util.test_mock'
This function was added in the first place because there needed to be a way to add mocks
when transitioning from executing a command from a file in addition to as a function. However,
since we add the mock when running the test (and BATS automatically isolates the added mocks in
each test), we don't have to re-setup the mocks when executing a particular function. Also,
environment variables were originally used because those (as opposed to Bash functions) are
copied to the child exec environment variable table