fix: Remove `cellar` folder. Closes #7
Given the instructions in the README, users will now clone neobasher
to ~/.local/share/neobasher/source. And because neobasher will automatically
use ~/.local/share/neobasher/cellar for it's cellar/NEOBASHER_PREFIX, the cellar
directory is extraneous, especially since it is not used. Commands already
call `mkdir -p` on various subfolders of cellar, so this is OK