fix: Remove erroneous 'Skipping fileName existing symlink' errors for man pages
This arised because the man page autodetector (search_heuristics function) searched subdirectories of subdirectories,
as well as directories. So when the root path was searched, it looked in the 'man' subdirectory. Since we already looked
in 'man', there was a collision. So the code was updated to only look for man _files_ at the root of the directory