fix: Remove check before 'unlink'. Closes #61
The check was originally implemented because it was required for
'abstract-mans' (and copied to the other abstract-*.sh files). This is
because the heuristics for the man linking implemented the same
search functionality for the root directory and the 'man' subdirectory,
which caused files to be detected twice. On the second unlink, a failure
occured. Since that code no longer exists, the check is no longer needed