refactor: Rename 'lib' directory to 'src'
'src' makes a bit more sense since non-libraries won't have their code
in a 'lib' directory. Furthermore, this 'lib' is unrelated to the
'/usr/lib' directory from the Filesystem Hierarchy Standard, so changing
the name makes that more explicit. This commit makes a bunch of other
changes to some other files and stuff that are sort of unrelated to the
main refactor